:recycle: Do an immediate rebuild upon block prediction

+++ Assets/Scripts/Events/Chunk/BlockPredictionEvent.cs
@@ -3,9 +3,9 @@ namespace CraftSharp.Event
     public record BlockPredictionEvent : BaseEvent
     {
         public BlockLoc BlockLoc { get; }
-        public ushort BlockStateId { get; }
+        public int BlockStateId { get; }
 
-        public BlockPredictionEvent(BlockLoc blockLoc, ushort blockStateId)
+        public BlockPredictionEvent(BlockLoc blockLoc, int blockStateId)
         {
             BlockLoc = blockLoc;
             BlockStateId = blockStateId;
